added controllers and finance tab for clients view page

+++ database/migrations/account/2023_10_24_115045_create_account_transactions_table.php
@@ -29,7 +29,7 @@ class CreateAccountTransactionsTable extends Migration
             $table->string('reason')->nullable();
             $table->date('reason_date')->nullable();
 
-            $table->unsignedSmallInteger('committer_id')->nullable();
+            $table->unsignedInteger('committer_id')->nullable();
 
             $table->text('comments')->nullable();
